+++ src/aux-records/AuthStore.ts
@@ -481,8 +481,9 @@ export interface AuthUser {
  * - "none" means that the user has no special permissions.
  * - "superUser" means that the user has additional permissions that only special users should have.
  * - "system" means that the user is the system and is performing a system operation.
+ * - "moderator" means that the user is a moderator and has additional permissions to moderate content.
  */
-export type UserRole = 'none' | 'superUser' | 'system';
+export type UserRole = 'none' | 'superUser' | 'system' | 'moderator';
 
 export interface AuthUserAuthenticator {
     /**