+++ src/aux-records/MemoryStore.ts
@@ -85,6 +85,7 @@ import {
     ResourcePermissionAssignment,
     RoleAssignment,
     UpdateUserRolesResult,
+    UserPrivacyFeatures,
     getExpireTime,
     getSubjectUserId,
 } from './PolicyStore';
@@ -806,7 +807,7 @@ export class MemoryStore
         this._studioHumeConfigs.set(studioId, config);
     }
 
-    async getUserPrivacyFeatures(userId: string): Promise<PrivacyFeatures> {
+    async getUserPrivacyFeatures(userId: string): Promise<UserPrivacyFeatures> {
         return await this._getUserPrivacyFeatures(userId);
     }
 
@@ -2273,10 +2274,13 @@ export class MemoryStore
 
     private async _getUserPrivacyFeatures(
         userId: string
-    ): Promise<PrivacyFeatures> {
+    ): Promise<UserPrivacyFeatures> {
         const user = await this.findUser(userId);
         if (user?.privacyFeatures) {
-            return user.privacyFeatures;
+            return {
+                ...user.privacyFeatures,
+                userRole: user?.role,
+            };
         }
 
         return {
@@ -2284,6 +2288,7 @@ export class MemoryStore
             allowPublicData: true,
             allowAI: true,
             allowPublicInsts: true,
+            userRole: user?.role,
         };
     }