+++ src/aux-records/PolicyController.spec.ts
@@ -2541,6 +2541,13 @@ describe('PolicyController', () => {
             ['listSubscriptions', 'resourceId'],
         ];
 
+        const moderatorActionCases: [ActionKinds][] = [
+            ['read'],
+            ['list'],
+            ['count'],
+            ['listSubscriptions'],
+        ];
+
         const adminOrGrantedResourceKindCases: [ResourceKinds][] = [
             ['data'],
             ['file'],
@@ -2796,6 +2803,56 @@ describe('PolicyController', () => {
                                 expireTimeMs: null,
                             },
                             explanation: 'Role is "admin".',
+                        });
+                    });
+
+                    it('should allow the action if the user is a superUser', async () => {
+                        const user = await store.findUser(userId);
+                        await store.saveUser({
+                            ...user,
+                            role: 'superUser',
+                        });
+
+                        const context =
+                            await controller.constructAuthorizationContext({
+                                recordKeyOrRecordName: recordName,
+                                userId: userId,
+                            });
+
+                        const result = await controller.authorizeSubject(
+                            context,
+                            {
+                                subjectId: userId,
+                                subjectType: 'user',
+                                resourceKind: resourceKind,
+                                action: action,
+                                resourceId: resourceId,
+                                markers: [marker],
+                            }
+                        );
+
+                        expect(result).toEqual({
+                            success: true,
+                            recordName: recordName,
+                            permission: {
+                                id: null,
+                                recordName,
+                                userId: null,
+
+                                // The role that record owners recieve
+                                subjectType: 'role',
+                                subjectId: ADMIN_ROLE_NAME,
+
+                                // resourceKind and action are null because this permission
+                                // applies to all resources and actions.
+                                resourceKind: null,
+                                action: null,
+
+                                marker: marker,
+                                options: {},
+                                expireTimeMs: null,
+                            },
+                            explanation: 'User is a superUser.',
                         });
                     });
 
@@ -3145,6 +3202,100 @@ describe('PolicyController', () => {
                             },
                         });
                     });
+
+                    if (moderatorActionCases.some(([a]) => action === a)) {
+                        it('should allow the action if the user is a moderator', async () => {
+                            const user = await store.findUser(userId);
+                            await store.saveUser({
+                                ...user,
+                                role: 'moderator',
+                            });
+
+                            const context =
+                                await controller.constructAuthorizationContext({
+                                    recordKeyOrRecordName: recordName,
+                                    userId: userId,
+                                });
+
+                            const result = await controller.authorizeSubject(
+                                context,
+                                {
+                                    subjectId: userId,
+                                    subjectType: 'user',
+                                    resourceKind: resourceKind,
+                                    action: action,
+                                    resourceId: resourceId,
+                                    markers: [marker],
+                                }
+                            );
+
+                            expect(result).toEqual({
+                                success: true,
+                                recordName: recordName,
+                                permission: {
+                                    id: null,
+                                    recordName,
+                                    userId: null,
+
+                                    // The role that record owners recieve
+                                    subjectType: 'role',
+                                    subjectId: ADMIN_ROLE_NAME,
+
+                                    // resourceKind and action are null because this permission
+                                    // applies to all resources and actions.
+                                    resourceKind: null,
+                                    action: action,
+
+                                    marker: marker,
+                                    options: {},
+                                    expireTimeMs: null,
+                                },
+                                explanation: 'User is a moderator.',
+                            });
+                        });
+                    } else {
+                        it('should deny the action even if the user is a moderator', async () => {
+                            const user = await store.findUser(userId);
+                            await store.saveUser({
+                                ...user,
+                                role: 'moderator',
+                            });
+
+                            const context =
+                                await controller.constructAuthorizationContext({
+                                    recordKeyOrRecordName: recordName,
+                                    userId: userId,
+                                });
+
+                            const result = await controller.authorizeSubject(
+                                context,
+                                {
+                                    subjectId: userId,
+                                    subjectType: 'user',
+                                    resourceKind: resourceKind,
+                                    action: action,
+                                    resourceId: resourceId,
+                                    markers: [marker],
+                                }
+                            );
+
+                            expect(result).toEqual({
+                                success: false,
+                                errorCode: 'not_authorized',
+                                errorMessage:
+                                    'You are not authorized to perform this action.',
+                                reason: {
+                                    type: 'missing_permission',
+                                    recordName: recordName,
+                                    subjectType: 'user',
+                                    subjectId: userId,
+                                    resourceKind: resourceKind,
+                                    action: action,
+                                    resourceId: resourceId,
+                                },
+                            });
+                        });
+                    }
                 }
             );
         });
@@ -4358,6 +4509,7 @@ describe('PolicyController', () => {
                                 publishData: false,
                             },
                             sendNotLoggedIn: true,
+                            userRole: 'none',
                         },
                     });